Added a speed up button for evolution purposes (Remove when handing in)
Robbit the Rabbit now included in Petlist. He has the special ability to evolve into any basic.

Evolution should work when time has passed (Obviously I'm not going to check real time so a button is used instead to check triggering)

When creating a new save, a random basic pet is generated (Uninstall and install App to check this)